Rename pecos.guppy module to pecos.guppy_gen
