+++ python/quantum-pecos/src/pecos/guppy_gen/transversal.py
