+++ python/quantum-pecos/tests/guppy/test_surface_ancilla_budget.py
@@ -7,7 +7,7 @@ import pytest
 
 def test_surface_qubit_count_respects_ancilla_budget() -> None:
     """The optional budget caps peak live ancillas without changing data qubits."""
-    from pecos.guppy import get_num_qubits
+    from pecos.guppy_gen import get_num_qubits
 
     assert get_num_qubits(7) == 97
     assert get_num_qubits(9) == 161
@@ -17,7 +17,7 @@ def test_surface_qubit_count_respects_ancilla_budget() -> None:
 
 def test_surface_ancilla_budget_must_be_positive() -> None:
     """Reject nonsensical budgets early."""
-    from pecos.guppy import get_num_qubits
+    from pecos.guppy_gen import get_num_qubits
 
     with pytest.raises(ValueError, match="ancilla_budget must be >= 1"):
         get_num_qubits(3, ancilla_budget=0)
@@ -26,7 +26,7 @@ def test_surface_ancilla_budget_must_be_positive() -> None:
 def test_constrained_ancilla_surface_code_compiles_to_hugr() -> None:
     """A budgeted surface memory experiment should still be valid Guppy/HUGR."""
     from pecos.compilation_pipeline import compile_guppy_to_hugr
-    from pecos.guppy import make_surface_code
+    from pecos.guppy_gen import make_surface_code
 
     program = make_surface_code(distance=3, num_rounds=1, basis="Z", ancilla_budget=2)
     hugr = compile_guppy_to_hugr(program)
