+++ python/quantum-pecos/tests/qec/surface/test_check_plan.py
@@ -222,7 +222,7 @@ def test_check_plan_and_interaction_basis_mismatch_fails_loudly() -> None:
 
 
 def test_guppy_surface_code_module_records_resolved_check_plan() -> None:
-    from pecos.guppy import get_surface_code_module
+    from pecos.guppy_gen import get_surface_code_module
 
     module = get_surface_code_module(3, check_plan="szz_current_v1")
 
@@ -233,7 +233,7 @@ def test_guppy_surface_code_module_records_resolved_check_plan() -> None:
 
 
 def test_guppy_surface_code_rejects_plan_basis_mismatch() -> None:
-    from pecos.guppy import make_surface_code
+    from pecos.guppy_gen import make_surface_code
 
     with pytest.raises(ValueError, match="conflicts with check_plan"):
         make_surface_code(
@@ -246,7 +246,7 @@ def test_guppy_surface_code_rejects_plan_basis_mismatch() -> None:
 
 
 def test_guppy_surface_code_accepts_check_plan_as_source_of_truth() -> None:
-    from pecos.guppy import make_surface_code
+    from pecos.guppy_gen import make_surface_code
 
     program = make_surface_code(
         distance=3,
@@ -269,7 +269,7 @@ def test_guppy_surface_code_accepts_check_plan_as_source_of_truth() -> None:
     ],
 )
 def test_guppy_surface_code_accepts_balanced_data_check_plans(check_plan: str) -> None:
-    from pecos.guppy import make_surface_code
+    from pecos.guppy_gen import make_surface_code
 
     program = make_surface_code(
         distance=3,
@@ -395,7 +395,7 @@ def test_direct_surface_renderers_accept_check_plan_as_source_of_truth() -> None
 
 
 def test_szz_guppy_source_can_disable_trace_metadata_for_execution() -> None:
-    from pecos.guppy.surface import generate_guppy_source
+    from pecos.guppy_gen.surface import generate_guppy_source
     from pecos.qec.surface import SurfacePatch
 
     patch = SurfacePatch.create(distance=3)
@@ -414,7 +414,7 @@ def test_szz_guppy_source_can_disable_trace_metadata_for_execution() -> None:
 
 
 def test_szz_runtime_barrier_fences_data_prefix_before_host() -> None:
-    from pecos.guppy.surface import generate_guppy_source
+    from pecos.guppy_gen.surface import generate_guppy_source
     from pecos.qec.surface import SurfacePatch
 
     patch = SurfacePatch.create(distance=3)
@@ -438,7 +438,7 @@ def test_szz_runtime_barrier_fences_data_prefix_before_host() -> None:
 
 
 def test_szz_data_prefixes_emit_generic_hosted_metadata() -> None:
-    from pecos.guppy.surface import generate_guppy_source
+    from pecos.guppy_gen.surface import generate_guppy_source
     from pecos.qec.surface import SurfacePatch
 
     source = generate_guppy_source(
@@ -463,7 +463,7 @@ def test_szz_data_prefixes_emit_generic_hosted_metadata() -> None:
 
 
 def test_szz_hosted_metadata_labels_include_helper_scope() -> None:
-    from pecos.guppy.surface import generate_guppy_source
+    from pecos.guppy_gen.surface import generate_guppy_source
     from pecos.qec.surface import SurfacePatch
 
     source = generate_guppy_source(
@@ -477,7 +477,7 @@ def test_szz_hosted_metadata_labels_include_helper_scope() -> None:
 
 
 def test_plain_szz_memory_source_unrolls_hosted_metadata_by_counted_round() -> None:
-    from pecos.guppy.surface import generate_guppy_source
+    from pecos.guppy_gen.surface import generate_guppy_source
     from pecos.qec.surface import SurfacePatch
 
     source = generate_guppy_source(
@@ -496,7 +496,7 @@ def test_plain_szz_memory_source_unrolls_hosted_metadata_by_counted_round() -> N
 
 
 def test_plain_szz_memory_cache_key_includes_counted_rounds() -> None:
-    from pecos.guppy.surface import _guppy_module_cache_key
+    from pecos.guppy_gen.surface import _guppy_module_cache_key
     from pecos.qec.surface import SurfacePatch
 
     patch = SurfacePatch.create(distance=3)
@@ -599,7 +599,7 @@ def test_round_order_szz_check_plan_changes_host_order_not_metadata(
     round_order_plan: str,
     expected_hosts: list[str],
 ) -> None:
-    from pecos.guppy.surface import generate_guppy_source
+    from pecos.guppy_gen.surface import generate_guppy_source
     from pecos.qec.surface import SurfacePatch
     from pecos.qec.surface.circuit_builder import (
         OpType,
